requirements.txt
fastmcp>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
asyncio
redis>=5.0.0
google-generativeai>=0.7.2
python-socketio>=5.0.0
PyJWT>=2.8.0
